agent_spec_v2/requirements.txt
# ── Web service ───────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.0

# ── LLM backends ──────────────────────────────────────────────────────────────
# Ollama (local models)
ollama>=0.3.0
# LM Studio / any OpenAI-compatible endpoint
openai>=1.30.0

# ── Code analysis — used by agent_spec_v2 tools via agent_spec modules ────────
# BM25 lexical search (bm25_search tool)
rank-bm25>=0.2.2

# Tree-sitter AST parsing (ast_analyse tool)
tree-sitter>=0.22
tree-sitter-python>=0.23
tree-sitter-javascript>=0.23
tree-sitter-typescript>=0.23
tree-sitter-java>=0.23
tree-sitter-go>=0.23
tree-sitter-rust>=0.23
tree-sitter-cpp>=0.23
tree-sitter-c>=0.23
tree-sitter-c-sharp>=0.23
tree-sitter-ruby>=0.23
tree-sitter-kotlin>=0.3

# Call graph & import graph (ast_analyse, import_graph tools)
networkx>=3.3

# Semantic search — ONNX-based, no GPU required (semantic_search tool)
fastembed>=0.3.0

# ── Utilities ─────────────────────────────────────────────────────────────────
httpx>=0.27.0
typing-extensions>=4.11.0
